Updates README and adds Dockerfile
The README describes how to run it locally and that is great if you have Ruby
installed. If you don't, never fear, Docker is here. The instructions for that
are in the README as well.

The Dockerfile moves the code into the file and then creates an entry point
and a file share. It's best not to run it with the long convulted command but
instead alias it.

Signed-off-by: Franklin Webber <franklin@chef.io>