Updates the Docker usage path
Alright, setting aliases means that the Bash subshell to evaluate the
current working directory was going to always been wherever you defined
the alias. So it was locked to one and only one directory.

This fixes it so that it will work in any diretory.

This also fixes the code to add it to all shells that are launched.

Signed-off-by: Franklin Webber <franklin@chef.io>